Generate gql for collection and noteslist queries
Refactor CollectionView to MyNotesContent.
