--- components/my-notes/collection-view/CollectionView.mocks.ts
@@ -1,5 +1,0 @@
-const base = {};
-
-export const mockCollectionViewProps = {
-  base,
-};
