+++ components/my-notes/my-notes-content/MyNotesContent.mocks.ts
@@ -1,0 +1,5 @@
+const base = {};
+
+export const mockMyNotesContentProps = {
+  base,
+};
