+++ components/my-notes/my-notes-content/MyNotesContent.stories.tsx
@@ -1,21 +1,21 @@
 import { ComponentMeta, ComponentStory } from "@storybook/react";
 import { SelectedCollectionProvider } from "../../../contexts/SelectedCollectionContext";
 import { SelectedListProvider } from "../../../contexts/SelectedListContext";
-import CollectionView from "./CollectionView";
-import { mockCollectionViewProps } from "./CollectionView.mocks";
+import MyNotesContent from "./MyNotesContent";
+import { mockMyNotesContentProps } from "./MyNotesContent.mocks";
 
 export default {
-  title: "my-notes/collection-view/CollectionView",
-  component: CollectionView,
+  title: "my-notes/my-notes-content/MyNotesContent",
+  component: MyNotesContent,
   // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
   argTypes: {},
-} as ComponentMeta<typeof CollectionView>;
+} as ComponentMeta<typeof MyNotesContent>;
 
 // More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
-const Template: ComponentStory<typeof CollectionView> = () => (
+const Template: ComponentStory<typeof MyNotesContent> = () => (
   <SelectedCollectionProvider>
     <SelectedListProvider>
-      <CollectionView />
+      <MyNotesContent />
     </SelectedListProvider>
   </SelectedCollectionProvider>
 );
@@ -24,5 +24,5 @@ export const Base = Template.bind({});
 // More on args: https://storybook.js.org/docs/react/writing-stories/args
 
 Base.args = {
-  ...mockCollectionViewProps.base,
+  ...mockMyNotesContentProps.base,
 };
