+++ components/my-notes/my-notes-content/left-pane-content/LeftPaneContent.stories.tsx
@@ -1,20 +1,22 @@
 import { ComponentMeta, ComponentStory } from "@storybook/react";
-import NoteContent from "./NoteContent";
-import { mockNoteContentProps } from "./NoteContent.mocks";
+import LeftPaneContent from "./LeftPaneContent";
+import { mockCollectionsProps } from "./LeftPaneContent.mocks";
 
 export default {
-  title: "my-notes/collection-view/note-content/NoteContent",
-  component: NoteContent,
+  title: "my-notes/my-notes-content/left-pane-content/LeftPaneContent",
+  component: LeftPaneContent,
   // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
   argTypes: {},
-} as ComponentMeta<typeof NoteContent>;
+} as ComponentMeta<typeof LeftPaneContent>;
 
 // More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
-const Template: ComponentStory<typeof NoteContent> = () => <NoteContent />;
+const Template: ComponentStory<typeof LeftPaneContent> = () => (
+  <LeftPaneContent />
+);
 
 export const Base = Template.bind({});
 // More on args: https://storybook.js.org/docs/react/writing-stories/args
 
 Base.args = {
-  ...mockNoteContentProps.base,
+  ...mockCollectionsProps.base,
 };
