+++ components/my-notes/my-notes-content/left-pane-content/LeftPaneContent.tsx
