+++ components/my-notes/my-notes-content/lists/Lists.stories.tsx
@@ -1,18 +1,16 @@
 import { ComponentMeta, ComponentStory } from "@storybook/react";
-import RightPaneContent from "./RightPaneContent";
-import { mockListsProps } from "./RightPaneContent.mocks";
+import Lists from "./Lists";
+import { mockListsProps } from "./Lists.mocks";
 
 export default {
-  title: "my-notes/collection-view/RightPaneContent",
-  component: RightPaneContent,
+  title: "my-notes/my-notes-content/lists/Lists",
+  component: Lists,
   // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
   argTypes: {},
-} as ComponentMeta<typeof RightPaneContent>;
+} as ComponentMeta<typeof Lists>;
 
 // More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
-const Template: ComponentStory<typeof RightPaneContent> = () => (
-  <RightPaneContent />
-);
+const Template: ComponentStory<typeof Lists> = () => <Lists />;
 
 export const Base = Template.bind({});
 // More on args: https://storybook.js.org/docs/react/writing-stories/args
