+++ components/my-notes/my-notes-content/note-content/NoteContent.stories.tsx
@@ -1,20 +1,20 @@
 import { ComponentMeta, ComponentStory } from "@storybook/react";
-import Collections from "./Collections";
-import { mockCollectionsProps } from "./Collections.mocks";
+import NoteContent from "./NoteContent";
+import { mockNoteContentProps } from "./NoteContent.mocks";
 
 export default {
-  title: "my-notes/collection-view/collections/Collections",
-  component: Collections,
+  title: "my-notes/my-notes-content/note-content/NoteContent",
+  component: NoteContent,
   // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
   argTypes: {},
-} as ComponentMeta<typeof Collections>;
+} as ComponentMeta<typeof NoteContent>;
 
 // More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
-const Template: ComponentStory<typeof Collections> = () => <Collections />;
+const Template: ComponentStory<typeof NoteContent> = () => <NoteContent />;
 
 export const Base = Template.bind({});
 // More on args: https://storybook.js.org/docs/react/writing-stories/args
 
 Base.args = {
-  ...mockCollectionsProps.base,
+  ...mockNoteContentProps.base,
 };
