+++ components/my-notes/my-notes-content/note-content/NoteContent.tsx
