+++ components/my-notes/my-notes-content/note-content/note-editor/NoteEditor.stories.tsx
@@ -1,22 +1,22 @@
 import { ComponentMeta, ComponentStory } from "@storybook/react";
-import LeftPaneContent from "./LeftPaneContent";
-import { mockCollectionsProps } from "./LeftPaneContent.mocks";
+import NoteEditor, { Props } from "./NoteEditor";
+import { mockNoteEditorProps } from "./NoteEditor.mocks";
 
 export default {
-  title: "my-notes/collection-view/left-pane-content/LeftPaneContent",
-  component: LeftPaneContent,
+  title: "my-notes/my-notes-content/note-content/note-editor/NoteEditor",
+  component: NoteEditor,
   // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
   argTypes: {},
-} as ComponentMeta<typeof LeftPaneContent>;
+} as ComponentMeta<typeof NoteEditor>;
 
 // More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
-const Template: ComponentStory<typeof LeftPaneContent> = () => (
-  <LeftPaneContent />
+const Template: ComponentStory<typeof NoteEditor> = (args) => (
+  <NoteEditor {...args} />
 );
 
 export const Base = Template.bind({});
 // More on args: https://storybook.js.org/docs/react/writing-stories/args
 
 Base.args = {
-  ...mockCollectionsProps.base,
-};
+  ...mockNoteEditorProps.base,
+} as Props;
