+++ components/my-notes/my-notes-content/note-content/note-editor/NoteEditor.tsx
