+++ components/my-notes/my-notes-content/right-pane-content/RightPaneContent.stories.tsx
@@ -1,22 +1,22 @@
 import { ComponentMeta, ComponentStory } from "@storybook/react";
-import NoteEditor, { Props } from "./NoteEditor";
-import { mockNoteEditorProps } from "./NoteEditor.mocks";
+import RightPaneContent from "./RightPaneContent";
+import { mockListsProps } from "./RightPaneContent.mocks";
 
 export default {
-  title: "my-notes/collection-view/note-content/note-editor/NoteEditor",
-  component: NoteEditor,
+  title: "my-notes/my-notes-content/RightPaneContent",
+  component: RightPaneContent,
   // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
   argTypes: {},
-} as ComponentMeta<typeof NoteEditor>;
+} as ComponentMeta<typeof RightPaneContent>;
 
 // More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
-const Template: ComponentStory<typeof NoteEditor> = (args) => (
-  <NoteEditor {...args} />
+const Template: ComponentStory<typeof RightPaneContent> = () => (
+  <RightPaneContent />
 );
 
 export const Base = Template.bind({});
 // More on args: https://storybook.js.org/docs/react/writing-stories/args
 
 Base.args = {
-  ...mockNoteEditorProps.base,
-} as Props;
+  ...mockListsProps.base,
+};
