+++ components/my-notes/my-notes-content/right-pane-content/RightPaneContent.tsx
