+++ graphql/queries/collections.graphql
@@ -1,0 +1,18 @@
+query collections {
+  collections {
+    id
+    title
+    visibility
+    upvotes
+    createdAt
+    updatedAt
+    owner {
+      id
+      username
+    }
+    lists {
+      id
+      title
+    }
+  }
+}
