+++ package.json
@@ -80,6 +80,7 @@
     "jest": "^28.1.1",
     "jest-environment-jsdom": "^28.1.1",
     "prettier": "^2.7.1",
+    "ts-node": "^10.9.1",
     "typescript": "4.7.4"
   },
   "resolutions": {
